Move theming to utilities folder
