+++ frontend/src/components/embeds/BleskEmbed.vue
@@ -1,5 +1,5 @@
 <script setup lang="ts">
-import ThemeProvider from '@/components/ThemeProvider.vue';
+import ThemeProvider from '@/components/utilities/theming/ThemeProvider.vue';
 </script>
 
 <template>
